Changed logic to keep fetches in same component
