+++ code/src/pages/CompanyList.js
@@ -6,12 +6,41 @@ import Error from '../components/Error';
 import { API_KEY } from '../api.js';
 import CompanyHeader from 'components/CompanyHeader';
 import CompanyDetail from 'components/CompanyDetail';
+import Backbutton from '../components/Backbutton';
+import MovieThumb from '../components/MovieThumb';
 
 const CompanyList = () => {
-  const [companyMovies, setCompanyMovies] = useState([]);
+  const [company, setCompany] = useState([]);
   const { companyId } = useParams();
   //const MOVIE_URL = `https://api.themoviedb.org/3/discover/movie?api_key=${API_KEY}&language=en-US&sort_by=popularity.desc&include_adult=false&include_video=false&page=1&with_companies=${companyId}`;
   const COMPANY_URL = `https://api.themoviedb.org/3/company/${companyId}?api_key=${API_KEY}`;
+
+  const [companyMovies, setCompanyMovies] = useState([]);
+  //const { companyId } = useParams();
+  const MOVIE_URL = `https://api.themoviedb.org/3/discover/movie?api_key=${API_KEY}&language=en-US&sort_by=popularity.desc&include_adult=false&include_video=false&page=1&with_companies=${companyId}`;
+
+  useEffect(() => {
+    fetch(MOVIE_URL)
+      .then(result => {
+        if (result.ok) {
+          return result.json();
+        } else {
+          throw new Error('404');
+        }
+      })
+      .then(json => {
+        console.log(json);
+        const filteredArray = json.results.filter(
+          movie => movie.poster_path != null
+        );
+
+        setCompanyMovies(filteredArray);
+        console.log(filteredArray);
+      })
+      .catch(() => {
+        window.location.assign('/error');
+      });
+  }, [MOVIE_URL]);
 
   // useEffect(() => {
   //   fetch(MOVIE_URL)
@@ -48,7 +77,7 @@ const CompanyList = () => {
       })
       .then(json => {
         console.log(json);
-        setCompanyMovies(json);
+        setCompany(json);
       })
       .catch(() => {
         window.location.assign('/error');
@@ -84,9 +113,17 @@ const CompanyList = () => {
   // };
   return (
     <>
-      <CompanyHeader {...companyMovies} />
+      <CompanyHeader {...company} />
       <main className="list">
-        <CompanyDetail />
+        <>
+          <Backbutton text={'Go Back!'} />
+          <div>
+            {companyMovies.map(company => (
+              <MovieThumb key={company.id} {...company} />
+            ))}
+          </div>
+        </>
+        {/* <CompanyDetail /> */}
         {/* {isLoading && <Loader />}
       {!isLoading && (
         <>
